refactor: ticket side panel design
