+++ desk/src/components/TicketField.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="flex gap-2 px-5 pb-1 leading-5 first:mt-3 items-center">
+  <div class="flex gap-2 pb-1 leading-5 items-center">
     <div class="w-[106px] shrink-0 truncate text-sm text-gray-600">
       <Tooltip :text="field.label">
         <span>{{ field.label }}</span>
